basic data model
For the time being, ember-data is not mature enough to be used with a
weird api with no real convention (like the one I'm using for this
blog).
So I created basic objects to emulate a simple version of ember model
lifecycle.
At this commit, it is read only, I haven't yet implemented update/saving
stuff.